requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
args==0.1.0
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==5.0.1
attrs==24.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.12.3
bounded-pool-executor==0.0.3
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1724954807150/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
clint==0.5.1
cmake==3.25.0
cnocr==2.3.0.3
cnstd==1.2.5.1
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1727293517607/work
coverage==7.6.8
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
Cython==3.0.11
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1732236634532/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docker-pycreds==0.4.0
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastprogress==1.0.3
filelock @ file:///croot/filelock_1700591183607/work
fire==0.7.0
flatbuffers==24.3.25
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1731643435391/work
freetype-py @ file:///home/conda/feedstock_root/build_artifacts/freetype-py_1650983368720/work
frozenlist==1.5.0
fsspec==2024.10.0
fvcore==0.1.5.post20221221
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1726922189413/work
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.36.0
imgaug==0.4.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1725921340658/work
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1731317204262/work
Jinja2 @ file:///croot/jinja2_1730902924303/work
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266648/work
lazy_loader==0.4
lightning-utilities==0.11.9
lit==15.0.7
lmdb==1.5.1
lxml==5.3.0
mamba==0.11.3
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///croot/networkx_1717597493534/work
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1731802491770/work
paddleocr==2.9.1
paddlepaddle-gpu==3.0.0b2
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1726878406085/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///croot/pillow_1731594689713/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
portalocker==3.0.0
pqdm==0.2.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.0
protobuf==5.29.0
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1729847094652/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
pycairo==1.27.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pydantic==2.10.2
pydantic_core==2.27.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1728880423364/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1731919281354/work
python-docx==1.1.2
pytorch-lightning==2.4.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1725430390447/work
RapidFuzz==3.10.1
rapidocr-onnxruntime==1.3.25
rdkit @ file:///home/conda/feedstock_root/build_artifacts/rdkit_1721395859551/work
reportlab @ file:///home/conda/feedstock_root/build_artifacts/reportlab_1727846838921/work
requests @ file:///croot/requests_1730999120400/work
rlPyCairo @ file:///home/conda/feedstock_root/build_artifacts/rlpycairo_1687519531733/work
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.19.0
setproctitle==1.3.4
shapely==2.0.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.5
SmilesPE==0.0.3
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1729066347171/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tomli==2.2.1
torch==2.0.1+cu118
torchaudio==2.0.2
torchmetrics==1.6.0
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm==4.67.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==2.0.0
typing_extensions==4.12.2
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1727140567071/work
ultralytics==8.3.39
ultralytics-thop==2.0.12
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1729704563364/work
Unidecode==1.3.8
urllib3 @ file:///croot/urllib3_1727769808118/work
wandb==0.18.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
wrapt==1.17.0
yacs==0.1.8
yarl==1.18.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1731262100163/work
